STATS_CHANGED

+++ av1/decoder/decoder.c
@@ -676,7 +676,7 @@ int av1_get_raw_frame(AV1Decoder *pbi, size_t index, YV12_BUFFER_CONFIG **sd,
 int av1_get_frame_to_show(AV1Decoder *pbi, YV12_BUFFER_CONFIG *frame) {
   if (pbi->num_output_frames == 0) return -1;
 #if CONFIG_OUTPUT_FRAME_BASED_ON_ORDER_HINT
-  const int out_frame_idx =
+  const size_t out_frame_idx =
       (pbi->common.seq_params.order_hint_info.enable_order_hint &&
        pbi->common.seq_params.enable_frame_output_order)
           ? 0
